requirements.txt
Flask==2.3.3
firebase-admin==6.5.0
qrcode==7.4
Pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0